fix get_class_labels by casting one_hot to float, default dtype is long

+++ dual_diffusion_pipeline.py
@@ -366,7 +366,7 @@ class DualDiffusionPipeline(DiffusionPipeline):
     @torch.no_grad()
     def get_class_labels(self, labels):
         if isinstance(labels, torch.Tensor):
-            return torch.nn.functional.one_hot(labels, num_classes=self.unet.label_dim).to(device=self.device)
+            return torch.nn.functional.one_hot(labels, num_classes=self.unet.label_dim).to(device=self.device, dtype=torch.float32)
         elif isinstance(labels, list):
             class_labels = torch.zeros((1, self.unet.label_dim))
             return class_labels.index_fill_(1, torch.tensor(labels, device=self.device, dtype=torch.long), 1).to(device=self.device)
